rework ovals and ellipse to share common base class
