+++ bokehjs/src/lib/models/glyphs/index.ts
@@ -3,7 +3,9 @@ export {Annulus}      from "./annulus"
 export {Arc}          from "./arc"
 export {Bezier}       from "./bezier"
 export {Circle}       from "./circle"
+export {CenterRotatable} from "./center_rotatable"
 export {Ellipse}      from "./ellipse"
+export {EllipseOval}  from "./ellipse_oval"
 export {Glyph}        from "./glyph"
 export {HBar}         from "./hbar"
 export {HexTile}      from "./hex_tile"
