+++ bokehjs/test/core/hittest.ts
@@ -10,12 +10,6 @@ describe("hittest module", () => {
     expect(hittest.point_in_ellipse(2.1, 1, 0, 1.5, 1, 1, 1)).to.be.equal(false)
   })
 
-  it("should return false if (x,y) point is outside an oval, true if inside", () => {
-    expect(hittest.point_in_oval(2, 1, 0, 2, 2, 1, 1)).to.be.equal(true)
-    expect(hittest.point_in_oval(2, 1, Math.PI/4, 2, 2, 1, 1)).to.be.equal(true)
-    expect(hittest.point_in_oval(3, 1, 0, 2, 2, 1, 1)).to.be.equal(false)
-  })
-
   it("should return false if (x,y) point is outside a polygon, true if inside", () => {
     expect(hittest.point_in_poly(1.5, 5, [1,2,2,1], [4,5,8,9])).to.be.equal(true)
     expect(hittest.point_in_poly(1.01, 4, [1,2,2,1], [4,5,8,9])).to.be.equal(false)
